A little optimization
I have noticed that when there are lots of collection views inside a view, the load is really high. I have a view that use 7 list, each one is a Collection View. Inside of each item of the list I have a table that it is again a Collection View of rows.
When I render the parent view it takes to me almost 3 seconds to load all in a modern browser.
So I did this modification to the code. But maybe there is another way to optimizate. I have got the times that It takes to each block. And my modification is fast. But my view with lots of Collection Views continue being slow. Maybe you could optimize other lines of your code knowing this little optimization.

Like Oz Katz describes in Avoiding Common Backbone Pitfalls appending collection views each time causes the DOM to reflow (meaning that the browser has to recalculate the position and size of every element in the DOM tree)
http://ozkatz.github.io/avoiding-common-backbonejs-pitfalls.html

But I found a problem with IE. Maybe in IE is not recommended.
http://stackoverflow.com/questions/14080365/bad-performance-ie-using-documentfragment

I will research more.