Addressing backbone.collection.add() with model array and at.
Removing the _renderModelViews method since the add trigger never will pass
an array.

Came to the realization that the collection holds the models at their
proper index and for situations where at has been provided we can
consult the collection on where it should be positioned rather than using

the 'at' values directly.