Fixed empty list caption handling with add.
Added a removeEmptyListCaption function to do the lifting and with a check
in _insertAndRenderModelView we now clean up any caption that maybe present.
Also adjusted the tests, which looked for a render event to now look for
an add event on the collection... got me thinking that maybe we should
trigger something unique on the collectionView end, but I'm leaving it
for now.