Making use of backbone.viewOptions.
backbone.viewOptions is now used and included in the source for use
in setting and updating values stored in the view.  The setOption()
function has been carried over for backwards compatibility.

Updated the versions of backbone and underscore in the test library
since backbone.viewOptions requires at least v1.1 of backbone.

Updated the package versions as well as the dependencies listing on
the README.

Added a changelog entry.